requirements.txt
alembic==1.13.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==23.2.0
autopage==0.5.2
captum==0.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cliff==4.6.0
cloudpickle==3.0.0
cmaes==0.10.0
cmd2==2.4.3
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docker-pycreds==0.4.0
econml==0.15.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
googledrivedownloader==0.4
greenlet==3.0.3
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
importlib-metadata==4.13.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lightgbm==4.3.0
llvmlite==0.42.0
loguru==0.7.2
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib==3.6.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nbformat==5.10.4
nbstripout==0.7.1
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.9.0.80
optuna==3.0.6
packaging==24.0
pandas==1.5.3
parso==0.8.4
patsy==0.5.6
pbr==6.0.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prettytable==3.10.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pynndescent==0.5.12
pyparsing==3.1.2
pyperclip==1.8.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.8.1
seaborn==0.12.2
sentry-sdk==2.1.1
setproctitle==1.3.3
shap==0.43.0
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sparse==0.15.1
SQLAlchemy==2.0.29
stack-data==0.6.3
statsmodels==0.14.2
stevedore==5.2.0
sympy==1.12
threadpoolctl==3.4.0
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.13
widgetsnbextension==4.0.10
zipp==3.18.1
